feat: add noise filter
