+++ src/components/pages/pricing/hero/contact-form/index.js
@@ -1,0 +1,3 @@
+import ContactForm from './contact-form';
+
+export default ContactForm;
